Extract utility method
