+++ src/lib.rs
@@ -54,3 +54,4 @@ pub mod machine;
 pub mod observables;
 pub mod program;
 pub mod state;
+mod util;
