+++ src/state.rs
@@ -8,6 +8,8 @@ use std::ops::Add;
 use rand::distributions::{Distribution, Uniform};
 use rand::Rng;
 
+use super::util::marsaglia;
+
 /// This trait specifies what a spin is.
 pub trait Spin: Clone + Add<Self, Output = Self::MagnetizationType> {
     type MagnetizationType: Magnetization<SpinType = Self>;
@@ -200,21 +202,9 @@ impl Spin for HeisenbergSpin {
         HeisenbergSpin([0f64, 0f64, -1f64])
     }
 
-    /// Gerate a random Heisenberg spin using the Marsaglia method for sphere
-    /// point picking.
     fn rand<T: Rng>(rng: &mut T) -> Self {
-        loop {
-            let distribution = Uniform::new(-1.0, 1.0);
-            let x1 = distribution.sample(rng);
-            let x2 = distribution.sample(rng);
-            if x1 * x1 + x2 * x2 >= 1f64 {
-                continue;
-            }
-            let x = 2f64 * x1 * (1f64 - x1 * x1 - x2 * x2).sqrt();
-            let y = 2f64 * x2 * (1f64 - x1 * x1 - x2 * x2).sqrt();
-            let z = 1f64 - 2f64 * (x1 * x1 + x2 * x2);
-            return HeisenbergSpin([x, y, z]);
-        }
+        let (x, y, z) = marsaglia(rng);
+        return HeisenbergSpin([x, y, z]);
     }
 
     #[inline]
